# backend/specs/openapi.yaml
openapi: 3.1.0
info:
  title: NarrativeOS API
  version: 0.3.0
paths:
  /v1/creator/commercial-blueprint:
    post:
      tags:
      - creator
      summary: Create Commercial Blueprint
      operationId: create_commercial_blueprint_v1_creator_commercial_blueprint_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommercialBlueprintRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Create Commercial Blueprint V1 Creator Commercial
                  Blueprint Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/creator/dialogue/sessions:
    post:
      tags:
      - creator
      summary: Create Dialogue Session
      operationId: create_dialogue_session_v1_creator_dialogue_sessions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatorDialogueSessionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Create Dialogue Session V1 Creator Dialogue Sessions
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/creator/dialogue/sessions/{session_id}:
    get:
      tags:
      - creator
      summary: Get Dialogue Session
      operationId: get_dialogue_session_v1_creator_dialogue_sessions__session_id__get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Dialogue Session V1 Creator Dialogue Sessions  Session
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/creator/dialogue/sessions/{session_id}/turns:
    post:
      tags:
      - creator
      summary: Add Dialogue Turn
      operationId: add_dialogue_turn_v1_creator_dialogue_sessions__session_id__turns_post
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatorDialogueTurnRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Add Dialogue Turn V1 Creator Dialogue Sessions  Session
                  Id  Turns Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/account/snapshot:
    get:
      tags:
      - account
      summary: Account Snapshot
      operationId: account_snapshot_v1_account_snapshot_get
      parameters:
      - name: account_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Account Id
      - name: reader_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Reader Id
      - name: creator_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Creator Id
      - name: include_diagnostics
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Diagnostics
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Account Snapshot V1 Account Snapshot Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/account/merge/preview:
    post:
      tags:
      - account
      summary: Account Merge Preview
      operationId: account_merge_preview_v1_account_merge_preview_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountMergePreviewRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Account Merge Preview V1 Account Merge Preview Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/account/merge/confirm:
    post:
      tags:
      - account
      summary: Account Merge Confirm
      operationId: account_merge_confirm_v1_account_merge_confirm_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountMergeConfirmRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Account Merge Confirm V1 Account Merge Confirm Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/account/delete/confirm:
    post:
      tags:
      - account
      summary: Account Delete Confirm
      operationId: account_delete_confirm_v1_account_delete_confirm_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDeleteConfirmRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Account Delete Confirm V1 Account Delete Confirm Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AccountDeleteConfirmRequest:
      properties:
        confirmation:
          type: string
          title: Confirmation
      type: object
      required:
      - confirmation
      title: AccountDeleteConfirmRequest
    AccountMergeConfirmRequest:
      properties:
        guest_reader_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Guest Reader Id
        guest_creator_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Guest Creator Id
        resolution:
          type: string
          title: Resolution
          default: keep_all_latest_first
      type: object
      title: AccountMergeConfirmRequest
    AccountMergePreviewRequest:
      properties:
        guest_reader_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Guest Reader Id
        guest_creator_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Guest Creator Id
        include_diagnostics:
          type: boolean
          title: Include Diagnostics
          default: false
      type: object
      title: AccountMergePreviewRequest
...
    CommercialBlueprintRequest:
      properties:
        creator_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Creator Id
        pen_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Pen Name
        genre:
          anyOf:
          - type: string
          - type: 'null'
          title: Genre
        audience:
          anyOf:
          - type: string
          - type: 'null'
          title: Audience
        commercial_goal:
          anyOf:
          - type: string
          - type: 'null'
          title: Commercial Goal
        platform:
          anyOf:
          - type: string
          - type: 'null'
          title: Platform
        tone:
          anyOf:
          - type: string
          - type: 'null'
          title: Tone
        seed:
          anyOf:
          - type: string
          - type: 'null'
          title: Seed
      type: object
      title: CommercialBlueprintRequest
...
    CreatorDialogueSessionRequest:
      properties:
        creator_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Creator Id
        seed:
          anyOf:
          - type: string
          - type: 'null'
          title: Seed
        tone:
          anyOf:
          - type: string
          - type: 'null'
          title: Tone
        target_length:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Length
        language:
          type: string
          title: Language
          default: zh-CN
        context:
          additionalProperties: true
          type: object
          title: Context
      type: object
      title: CreatorDialogueSessionRequest
    CreatorDialogueTurnRequest:
      properties:
        message:
          type: string
          title: Message
        context:
          additionalProperties: true
          type: object
          title: Context
        previous_session:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Previous Session
      type: object
      required:
      - message
      title: CreatorDialogueTurnRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...